requirements.txt
kafka-python==1.4.7



